+++ app/src/main/java/eu/kanade/tachiyomi/data/database/models/TrackImpl.kt
@@ -20,7 +20,7 @@ class TrackImpl : Track {
 
     override var score: Double = 0.0
 
-    override var status: Int = 0
+    override var status: Long = 0
 
     override var started_reading_date: Long = 0
