+++ app/src/main/java/eu/kanade/tachiyomi/data/track/Tracker.kt
@@ -27,15 +27,15 @@ interface Tracker {
     @DrawableRes
     fun getLogo(): Int
 
-    fun getStatusList(): List<Int>
+    fun getStatusList(): List<Long>
 
-    fun getStatus(status: Int): StringResource?
+    fun getStatus(status: Long): StringResource?
 
-    fun getReadingStatus(): Int
+    fun getReadingStatus(): Long
 
-    fun getRereadingStatus(): Int
+    fun getRereadingStatus(): Long
 
-    fun getCompletionStatus(): Int
+    fun getCompletionStatus(): Long
 
     fun getScoreList(): ImmutableList<String>
 
@@ -70,7 +70,7 @@ interface Tracker {
     // TODO: move this to an interactor, and update all trackers based on common data
     suspend fun register(item: Track, mangaId: Long)
 
-    suspend fun setRemoteStatus(track: Track, status: Int)
+    suspend fun setRemoteStatus(track: Track, status: Long)
 
     suspend fun setRemoteLastChapterRead(track: Track, chapterNumber: Int)
